Changed min/max value handling
The new settings allow to separately define minimum and maximum
number of occurrences of an argument in the input. That allows
definining required values (min count > 0), setting range of
allowed count (e.g. verbosity from 1 to 3 of "-v" occurrences) and
creating more complicated setup of positional arguments (e.g. first
two are "patterns", the rest are "files" and only the first two
required).

Tests and examples updated to use the new API.